+++ packages/atlas/src/providers/wallet/wallet.helpers.ts
@@ -1,4 +1,6 @@
-import { BaseDotsamaWallet, getWallets } from '@talismn/connect-wallets'
+import { BaseDotsamaWallet, Wallet, getWallets } from '@talismn/connect-wallets'
+
+import { atlasConfig } from '@/config'
 
 export const getWalletsList = () => {
   const supportedWallets = getWallets()
@@ -14,7 +16,21 @@ export const getWalletsList = () => {
     return [...acc, new UnknownWallet(walletName)]
   }, [] as UnknownWallet[])
 
-  return [...supportedWallets, ...unknownWallets]
+  const wcWallet = {
+    extensionName: 'WalletConnect',
+    title: 'WalletConnect',
+    logo: {
+      src: 'https://dev.gleev.xyz/favicon.ico',
+      alt: 'WalletConnect',
+    },
+    installed: true,
+  } as Wallet
+
+  return [
+    ...supportedWallets,
+    ...unknownWallets,
+    ...(atlasConfig.features.walletConnect.walletConnectProjectId ? [wcWallet] : []),
+  ]
 }
 
 // eslint-disable-next-line @typescript-eslint/no-explicit-any