+++ CHANGELOG.md
@@ -5,6 +5,20 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [5.2.1] - 2024-05-06
+
+### Added
+
+- VWO feature
+- WalletConnect integration
+- YPP Dashboard remake
+
+### Fixed
+
+- HTML tags fixed
+- YPP variation page
+- Small CRT fixes
+
 ## [5.2.0] - 2024-04-22
 
 ### Added